1.0.2 branch - fix the issue 48
svn path=/branches/1.0.2/; revision=323

+++ src/main/java/org/odlabs/wiquery/ui/slider/Slider.java
@@ -160,7 +160,7 @@ public class Slider extends WebMarkupContainer implements IWiQueryPlugin {
 	 */
 	public Number getMin() {
 		if(this.options.containsKey("min")){
-			return this.options.getInt("min");
+			return this.options.getFloat("min");
 		}
 		
 		return 0;
